Day 9
This one was pretty straightforward and I knew what to do pretty much from the beginning.  Sadly, I tried to use streams to make it "better" instead of a for loop and learned that stream().max() only dealt with positive numbers, causing all manner of headache.

I ended up trying out IntelliJ's new AI assistant, which is pretty good.  I did have to correct it a number of times, but it's definitely better than generic chatGPT for such things for sure.

Part 2 was < 10 minutes because recursion, that was nice.